Documents/time_series_project/stats_service/requirements.txt
fastapi
uvicorn[standard]
pydantic
numpy
sqlalchemy
psycopg2-binary
python-dotenv
